test: :white_check_mark: 单测修复&增加 useSpreadSheet 单测
